Enabled level 4 warnings and treating warnings as errors in MultiPLAY.vcxproj. Tweaked the code until it builds without any warnings. Fixed up MultiPLAY.vcxproj.filters using the IDE. Updated MultiPLAY.vcxproj to point at the header files with ClInclude elements. Created a new MultiPLAY.sln file. Sorted out the #include statements in MultiPLAY.cc. Most of them aren't needed now that the bulk of the code has been moved out to separate source code modules. Fixed it_envelope_node in Load_IT.cc to store the tick offset of each node using more than 8 bits. Added the Windows build output folder "x64" to .gitignore.

+++ conversion.cc
@@ -8,12 +8,12 @@ using namespace std;
 
 namespace MultiPLAY
 {
-	extern int from_lsb2(unsigned char in[2])
+	extern short from_lsb2(unsigned char in[2])
 	{
 		return in[0] | (static_cast<signed char>(in[1]) << 8);
 	}
 
-	extern unsigned int from_lsb2_u(unsigned char in[2])
+	extern unsigned short from_lsb2_u(unsigned char in[2])
 	{
 		return in[0] | (in[1] << 8);
 	}
@@ -33,12 +33,12 @@ namespace MultiPLAY
 		return in[0] | (in[1] << 8) | (in[2] << 16) | (in[3] << 24);
 	}
 
-	extern int from_msb2(unsigned char in[2])
+	extern short from_msb2(unsigned char in[2])
 	{
 		return in[1] | (static_cast<signed char>(in[0]) << 8);
 	}
 
-	extern unsigned int from_msb2_u(unsigned char in[2])
+	extern unsigned short from_msb2_u(unsigned char in[2])
 	{
 		return in[1] | (in[0] << 8);
 	}
